Added WPILib to SVN so I could easily diff the changes. :)
git-svn-id: https://kwarqs-dev.bbn.com/svn/kwarqs/code@61 eb211378-2a44-443b-ae4a-d45323d45073

+++ WPILib/PIDSource.h
@@ -15,7 +15,7 @@
 class PIDSource
 {
 public:
-	virtual double PIDGet() = 0;
+	virtual float PIDGet() = 0;
 };
 
 #endif
